# Drops duplicate edges in non-MultiGraph PLC `SGGraph` instances